+++ src/literalizer/languages/racket.py
@@ -41,6 +41,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     LanguageCls,
     OrderedMapFormatConfig,
     PrefixCallStyle,
@@ -299,6 +300,11 @@ class Racket(metaclass=LanguageCls):
         ERROR = NO_HETEROGENEOUS_BEHAVIOR
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.KEBAB,
+        IdentifierCase.SNAKE,
+    )
 
     validate_spec_for_data = no_validate_spec_for_data