+++ src/literalizer/languages/vb.py
@@ -53,6 +53,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     LanguageCls,
     OrderedMapFormatConfig,
     SequenceFormatConfig,
@@ -381,6 +382,11 @@ class VisualBasic(metaclass=LanguageCls):
         ERROR = NO_HETEROGENEOUS_BEHAVIOR
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.PASCAL,
+        IdentifierCase.CAMEL,
+    )
 
     validate_spec_for_data = no_validate_spec_for_data