+++ tests/integration/cases/call_ref_args_converted/Groovy_call.groovy
@@ -1,0 +1,13 @@
+def process(Map _args) { null }
+def myVar = [
+    1,
+    2,
+    3,
+]
+def myOther = [
+    4,
+    5,
+    6,
+]
+process(data: myVar, count: 42)
+process(data: myOther, count: 7)