+++ tests/integration/cases/call_ref_args_converted_whole/Groovy_call.groovy
@@ -1,0 +1,7 @@
+def process(Map _args) { null }
+def myVar = [
+    1,
+    2,
+    3,
+]
+process(data: myVar)