+++ tests/integration/cases/call_ref_args_converted_whole/Kotlin_call.kts
@@ -1,0 +1,7 @@
+fun process(data: Any? = null): Any? = null
+val myVar = intArrayOf(
+    1,
+    2,
+    3,
+)
+process(data = myVar)