+++ tests/integration/cases/call_ref_args_converted_whole/R_call.R
@@ -1,0 +1,6 @@
+my_var <- list(
+    1,
+    2,
+    3
+)
+process(data = my_var)